+++ txscript/stdscript/addressv0_test.go
