+++ txscript/stdscript/error.go
