+++ txscript/stdscript/error_test.go
