+++ txscript/stdscript/script.go
