+++ txscript/stdscript/script_bench_test.go
