+++ txscript/stdscript/script_test.go
