+++ txscript/stdscript/scriptshortform_test.go
