+++ txscript/stdscript/scriptv0.go
