+++ txscript/stdscript/scriptv0_test.go
